emotion-diffuser/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
python-multipart
python-dotenv
httpx
transformers
torch
datasets
openai
streamlit
pytest
